primary key type changes

+++ src/main/java/com/doctor/doctorregistration/models/DoctorDetails.java
@@ -12,7 +12,7 @@ import java.math.BigInteger;
 public class DoctorDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private BigInteger id;
     private String name;
     private String clinic_name;
     private String address;
@@ -90,11 +90,11 @@ public class DoctorDetails {
         this.password = password;
     }
 
-    public int getId() {
+    public BigInteger getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(BigInteger id) {
         this.id = id;
     }
 }
